Rewrite to support py3
